Adjust store access patterns
