+++ src/cpu/addressing.rs
@@ -293,71 +293,6 @@ mod index_indirect_x_addressing {
 }
 
 #[cfg(test)]
-mod indirect_index_y_addressing {
-  use crate::cpu::{
-    CPU,
-    addressing::indirect::IndirectIndexYAddressingTasks,
-    tests::{MemoryMock, run_tasks},
-  };
-
-  #[test]
-  fn should_return_address_offset_by_index_register_y_which_is_stored_at_zero_page_address() {
-    let mut memory = MemoryMock::new(&[0x02, 0xFF, 0x03, 0xDD, 0x25]);
-    let mut cpu = CPU::new_nmos();
-    cpu.index_register_y = 0x02;
-    cpu.program_counter = 0x00;
-
-    let mut tasks = Box::new(IndirectIndexYAddressingTasks::new());
-    run_tasks(&mut cpu, &mut *tasks, &mut memory);
-
-    assert_eq!(cpu.addr.value(), Some(0xDD05));
-  }
-
-  #[test]
-  fn should_advance_program_counter_once() {
-    let mut memory = MemoryMock::new(&[0x02, 0xFF, 0x03, 0xDD, 0x25]);
-    let mut cpu = CPU::new_nmos();
-    cpu.index_register_y = 0x02;
-    cpu.program_counter = 0x00;
-
-    let mut tasks = Box::new(IndirectIndexYAddressingTasks::new());
-    run_tasks(&mut cpu, &mut *tasks, &mut memory);
-
-    assert_eq!(cpu.program_counter, 0x01);
-  }
-
-  #[test]
-  fn should_take_four_cycles_when_not_crossing_page_boundary_during_offset_addition_for_a_read_operation_address()
-   {
-    let mut memory = MemoryMock::new(&[0x02, 0xFF, 0x03, 0xDD, 0x25]);
-    let mut cpu = CPU::new_nmos();
-    cpu.index_register_y = 0x02;
-    cpu.program_counter = 0x00;
-    cpu.cycle = 0;
-
-    let mut tasks = Box::new(IndirectIndexYAddressingTasks::new());
-    run_tasks(&mut cpu, &mut *tasks, &mut memory);
-
-    assert_eq!(cpu.cycle, 4);
-  }
-
-  #[test]
-  fn should_take_five_cycles_when_crossing_page_boundary_during_offset_addition_for_a_read_operation_address()
-   {
-    let mut memory = MemoryMock::new(&[0x02, 0xFF, 0x03, 0xDD, 0x25]);
-    let mut cpu = CPU::new_nmos();
-    cpu.index_register_y = 0xFF;
-    cpu.program_counter = 0x00;
-    cpu.cycle = 0;
-
-    let mut tasks = Box::new(IndirectIndexYAddressingTasks::new());
-    run_tasks(&mut cpu, &mut *tasks, &mut memory);
-
-    assert_eq!(cpu.cycle, 5);
-  }
-}
-
-#[cfg(test)]
 mod zero_page_addressing {
   use crate::cpu::{
     CPU,
